first login popup done
