--- src/app/login/layout.js
@@ -1,17 +1,0 @@
-import { Inter } from "next/font/google";
-import "../globals.css";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata = {
-  title: "zeit",
-  description: "online Stundenzettel Cloud App",
-};
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="de">
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
-}
