+++ src/app/page.js
@@ -1,7 +1,15 @@
+"use client";
+import { useState } from "react";
 import Image from "next/image";
 import styles from "./page.module.css";
+import Login from "./components/Login";
 
 export default function Home() {
+  const [seen, setSeen] = useState(false);
+
+  function togglePop() {
+    setSeen(!seen);
+  }
   return (
     <main className={styles.main}>
       <div className={styles.description}>
@@ -25,7 +33,7 @@ export default function Home() {
       </div>
 
       <div className={styles.center}>
-        <h1>zeit</h1>
+        <h1>z e i t</h1>
       </div>
 
       <div className={styles.grid}>
@@ -65,17 +73,13 @@ export default function Home() {
           <p>Watch a 5 min explains it all video</p>
         </a>
 
-        <a
-          href="https://google.de"
-          className={styles.card}
-          target="_blank"
-          rel="noopener noreferrer"
-        >
+        <button onClick={togglePop} className={styles.card}>
           <h2>
             Login <span>-&gt;</span>
           </h2>
           <p></p>
-        </a>
+        </button>
+        {seen ? <Login toggle={togglePop} /> : null}
       </div>
     </main>
   );
