+++ src/app/page.module.css
@@ -228,3 +228,88 @@
     transform: rotate(0deg);
   }
 }
+/* login popup */
+.popup {
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0, 0, 0, 0.4);
+}
+
+.popup_inner {
+  background-color: white;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  padding: 20px;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.75);
+  width: 40%;
+}
+
+.popup_inner h2 {
+  margin-top: 0;
+}
+
+.popup_inner label {
+  display: block;
+  margin-bottom: 10px;
+}
+
+.popup_inner input {
+  width: 100%;
+  padding: 5px;
+  border-radius: 5px;
+  border: 1px solid #ccc;
+  margin-bottom: 10px;
+}
+
+.popup_inner button[type="submit"],
+.popup_inner button[type="button"] {
+  background-color: #4caf50;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  margin-right: 10px;
+  cursor: pointer;
+  border-radius: 5px;
+}
+
+.popup_inner button[type="submit"]:hover,
+.popup_inner button[type="button"]:hover {
+  background-color: #45a049;
+}
+
+.popup_inner button[type="submit"] {
+  margin-top: 20px;
+}
+
+.popup_inner button[type="button"] {
+  margin-top: 20px;
+  background-color: #f44336;
+}
+
+.popup_inner button[type="button"]:hover {
+  background-color: #da190b;
+}
+
+.popup_inner button[type="submit"]:disabled {
+  background-color: #bfbfbf;
+  color: #ffffff;
+  cursor: not-allowed;
+}
+
+.popup_inner button[type="button"]:disabled {
+  background-color: #bfbfbf;
+  color: #ffffff;
+  cursor: not-allowed;
+}
